CLI, more logical operators, documentation, README.md (#8)
* simplified lexer, added strings

* evaluated strings

* small cleaning

* added a small test for newlines

* strings concatenation

* readme

* readme

* readme

* CLI, basic commands in console

* started to write a basic explanation of installation

* removed headline ipret

* added a template for readme

* version flag via build

* improved description of flags

* commentaries

* commentaries

* added operators `&&`, `||`, `>=` and `<=`, added more examples

* updated readme

* updated readme

* updated readme

* updated readme

* updated readme

* updated CHANGELOG.md

---------

Authored-by: DenisKozarezov <kozarezov.denis@rwb.ru>

+++ internal/parser/parser.go
@@ -10,7 +10,7 @@ import (
 
 type lexer interface {
 	NextToken() tokens.Token
-	CurrentLine() int64
+	CurrentLine() int16
 	CurrentPositionAtLine() int64
 }
